+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.2.0] - 2016-03-10
+### Added
+* First implementation of YasstFile encrypt/decrypt actions
+### Changed
+* Update README with gem release instructions
 
 ## [0.1.1] - 2016-02-11
 ### Changed
@@ -15,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Initial Release
 
-[Unreleased]: https://github.com/rdark/yasst/compare/0.1.1...develop
+[Unreleased]: https://github.com/rdark/yasst/compare/0.2.0...develop
+[0.2.0]: https://github.com/rdark/yasst/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/rdark/yasst/tree/0.1.0...0.1.1
 [0.1.0]: https://github.com/rdark/yasst/tree/0.1.0